+++ docs/api/changelog.rst
@@ -116,6 +116,8 @@ Version 0.9
 * Roster pushes are now accepted also if the :attr:`~.StanzaBase.from_` is the
   bare local JID instead of missing/empty (those are semantically equivalent).
 
+* :mod:`aioxmpp.hashes`
+
 .. _api-changelog-0.8:
 
 Version 0.8